requirements.txt
pdfminer.six==20191110
pkg-resources==0.0.0
pycryptodome==3.9.8
six==1.15.0
sortedcontainers==2.2.2
